Makes wakeup independent from result_queue
tomMoral/loky#48
This avoids deadlocks if a Process dies while:
* Unpickling the _CallItem
* Pickling a _ResultItem

Wakeups are done with _Sentinel object that cannot be used with wait.
We do not use a Connection/Queue as it brings lot of overhead in the
Executor to use only a small part of it. We might want to implement a
Sentinel object that can be waited upon to simplify and robustify the
code.

Test no deadlock with crashes
* TST crash in CallItem unpickling
* TST crash in func call run
* TST crash in REsult pickling
the test include crashes with PythonError/SystemExist/SegFault